py_server/requirements.txt
fastapi
uvicorn
uvicorn[standard]
mysql-connector-python
pandas
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl